refactor : InstagramText dto 구조 변경
