+++ TecheerPicture/src/main/java/com/techeerpicture/TecheerPicture/InstagramText/repository/InstagramTextRepository.java
@@ -1,4 +1,6 @@
-package com.techeerpicture.TecheerPicture.instagramtext;
+package com.techeerpicture.TecheerPicture.InstagramText.repository;
+
+import com.techeerpicture.TecheerPicture.InstagramText.entity.InstagramText;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
